# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simanja (Sistem Management Jamaah)",
        "description": "This is developped using Swagger UI.  You can find\r\nout more about Swagger at\r\n[http://swagger.io](http://swagger.io) or on\r\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "iqbalhanif313@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/setup/status-jamaah": {
            "get": {
                "tags": [
                    "setup/status-jamaah"
                ],
                "summary": "Show Setup StatusJamaah information",
                "operationId": "setup/status-jamaah",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "setup/status-jamaah"
                ],
                "summary": "Store Setup StatusJamaah",
                "operationId": "setup/status-jamaah/store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Store Setup StatusJamaah",
                                        "type": "string",
                                        "example": "SPR"
                                    },
                                    "nama": {
                                        "description": "Store Setup StatusJamaah",
                                        "type": "string",
                                        "example": "Semampir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/status-jamaah/{id}": {
            "get": {
                "tags": [
                    "setup/status-jamaah"
                ],
                "summary": "Show Detail Setup StatusJamaah",
                "operationId": "setup/status-jamaah/{id}",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "setup/status-jamaah"
                ],
                "summary": "Update Setup StatusJamaah",
                "operationId": "setup/status-jamaah/update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Update Setup StatusJamaah",
                                        "type": "string",
                                        "example": "SPR"
                                    },
                                    "nama": {
                                        "description": "Update Setup StatusJamaah",
                                        "type": "string",
                                        "example": "Semampir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "setup/status-jamaah"
                ],
                "summary": "Delete Setup StatusJamaah",
                "operationId": "setup/status-jamaah/{id}/delete",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}